#3 common模块实现config、common对象

+++ husky-entity/src/main/java/com/corkili/husky/common/Constants.java
@@ -2,7 +2,7 @@ package com.corkili.husky.common;
 
 public final class Constants {
 
-    public static final int EXISTED = 0;
-    public static final int DELETED = 1;
+    public static final byte EXISTED = 0;
+    public static final byte DELETED = 1;
 
 }
